# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LifeOS V1",
    "description": "LifeOS V1 Backend Architecture",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register",
        "description": "Registers a new user and returns their basic info.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_UserResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login",
        "description": "Authenticates a user and returns a JWT access token.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_TokenData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/execution/today": {
      "get": {
        "tags": [
          "Daily Execution"
        ],
        "summary": "Get Today",
        "description": "Fetch today's habits based on active plan.",
        "operationId": "get_today_api_v1_execution_today_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_List_TodaysHabitResponse__"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/execution/habit/complete": {
      "post": {
        "tags": [
          "Daily Execution"
        ],
        "summary": "Complete Habit",
        "description": "Mark a daily habit as completed.",
        "operationId": "complete_habit_api_v1_execution_habit_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogCompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_LogResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/stats/profile": {
      "get": {
        "tags": [
          "Statistics & Processing"
        ],
        "summary": "Get Profile",
        "description": "Get the current aggregated stats for the user's profile.",
        "operationId": "get_profile_api_v1_stats_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_UserStatResponse_"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/stats/process-day": {
      "post": {
        "tags": [
          "Statistics & Processing"
        ],
        "summary": "Process Day",
        "description": "Manual trigger to process today's logs for the authenticated user.\nSimulates the end of day evaluation.",
        "operationId": "process_day_api_v1_stats_process_day_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_Dict_str__Any__"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BaseResponse_Dict_str__Any__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "BaseResponse[Dict[str, Any]]"
      },
...
      "BaseResponse_List_TodaysHabitResponse__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TodaysHabitResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "BaseResponse[List[TodaysHabitResponse]]"
      },
      "BaseResponse_LogResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LogResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "BaseResponse[LogResponse]"
      },
...
      "BaseResponse_TokenData_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenData"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "BaseResponse[TokenData]"
      },
      "BaseResponse_UserResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "BaseResponse[UserResponse]"
      },
      "BaseResponse_UserStatResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserStatResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "BaseResponse[UserStatResponse]"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HabitCategory": {
        "type": "string",
        "enum": [
          "focus",
          "health",
          "discipline",
          "mind"
        ],
        "title": "HabitCategory"
      },
...
      "LogCompletionRequest": {
        "properties": {
          "habit_id": {
            "type": "integer",
            "title": "Habit Id"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "habit_id"
        ],
        "title": "LogCompletionRequest"
      },
      "LogResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "habit_name": {
            "type": "string",
            "title": "Habit Name"
          },
          "category": {
            "$ref": "#/components/schemas/HabitCategory"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "awarded_points": {
            "type": "integer",
            "title": "Awarded Points"
          },
          "late_flag": {
            "type": "boolean",
            "title": "Late Flag"
          }
        },
        "type": "object",
        "required": [
          "id",
          "habit_name",
          "category",
          "status",
          "awarded_points",
          "late_flag"
        ],
        "title": "LogResponse"
      },
...
      "TodaysHabitResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "habit_id": {
            "type": "integer",
            "title": "Habit Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "$ref": "#/components/schemas/HabitCategory"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          },
          "base_score": {
            "type": "integer",
            "title": "Base Score"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "completion_timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completion Timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "habit_id",
          "name",
          "category",
          "difficulty",
          "base_score",
          "status"
        ],
        "title": "TodaysHabitResponse"
      },
      "TokenData": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "TokenData"
      },
      "UserCreate": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "UTC"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UserCreate"
      },
      "UserLogin": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UserLogin"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "timezone"
        ],
        "title": "UserResponse"
      },
      "UserStatResponse": {
        "properties": {
          "total_points": {
            "type": "integer",
            "title": "Total Points"
          },
          "current_streak": {
            "type": "integer",
            "title": "Current Streak"
          },
          "max_streak": {
            "type": "integer",
            "title": "Max Streak"
          },
          "rank": {
            "type": "string",
            "title": "Rank"
          },
          "focus_points": {
            "type": "integer",
            "title": "Focus Points"
          },
          "health_points": {
            "type": "integer",
            "title": "Health Points"
          },
          "discipline_points": {
            "type": "integer",
            "title": "Discipline Points"
          },
          "mind_points": {
            "type": "integer",
            "title": "Mind Points"
          }
        },
        "type": "object",
        "required": [
          "total_points",
          "current_streak",
          "max_streak",
          "rank",
          "focus_points",
          "health_points",
          "discipline_points",
          "mind_points"
        ],
        "title": "UserStatResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/auth/login"
          }
        }
      }
    }
  }
}